aufgabe 5.2: sphere.cpp and sphere.h added (not tested yet)

+++ framework/shape.h
@@ -8,7 +8,7 @@
 class Shape{
 public:
     virtual double area()const = 0;
-    virtual float volume()const = 0;
+    virtual double volume()const = 0;
 };
 
 #endif //PROJECT_SHAPE_H
